MD5 and bugfixes
Added expected MD5 of file to patch and compare with final output.

Fixed a bug where we were ignoring some data on the longer file.

Ways to store MD5 as ints in a list to fit nicely with the other data.

Added support for command line arguments. mode file1 file2. file2 is ignored when using mode 2.